Refactor: man as notes.md Feature: added plugins ycm-core/YouCompleteMe, itchyny/lightline.vim, nathanaelkane/vim-indent-guides

+++ install.sh
@@ -12,14 +12,17 @@
 
 if (test -e ~/.vimrc_backup)
 then
+	echo "Making a backup of the previous backup folder"
 	mv ~/.vimrc_backup ~/.vimrc_backup_movein$(date +"%d%h%y_%H-%M-%S")
 fi
 
 if (test -e ~/.vimrc)
 then
+	echo "Making a backup of the previous vimrc file"
 	mv ~/.vimrc ~/.vimrc_backup
 fi
 
+echo "Installing the vimrc file"
 cp vimrc ~/.vimrc
 
 if (! test -d ~/.dic)
@@ -27,8 +30,16 @@ then
 	mkdir ~/.dic
 fi
 
-wget http://downloads.sourceforge.net/wordlist/hunspell-en_US-2020.12.07.zip
-unzip hunspell-en_US-2020.12.07.zip en_US.dic
-rm hunspell-en_US-2020.12.07.zip
-mv en_US.dic ~/.dic/
+if [ ! -f ~/.dic/en_US.dic ]; then
+	echo "Installing the English dictionary"
+	wget http://downloads.sourceforge.net/wordlist/hunspell-en_US-2020.12.07.zip
+	unzip hunspell-en_US-2020.12.07.zip en_US.dic
+	rm hunspell-en_US-2020.12.07.zip
+	mv en_US.dic ~/.dic/
+else
+	echo "The English dictionary already exists"
+fi
+
+echo "All Done."
+
